Allow list of locked players to be specified
Added a new config option.  You can provide a text file with player names in it.
The names in the file are locked.  So when you do a full roster overhaul you can
identify a set of players that need to be kept on the roster.  This is used in
addition to locking players based on their ownership percentage.